[martin-dev] two image comp
